move clearBlockFormat into clearFormat and deprecate
