+++ packages/roosterjs-editor-api/lib/index.ts
@@ -1,6 +1,6 @@
 export { default as changeFontSize, FONT_SIZES } from './format/changeFontSize';
 export { default as clearBlockFormat } from './format/clearBlockFormat';
-export { default as clearFormat, FormattingStrategy } from './format/clearFormat';
+export { default as clearFormat } from './format/clearFormat';
 export { default as createLink } from './format/createLink';
 export { default as getFormatState, getElementBasedFormatState } from './format/getFormatState';
 export { default as insertEntity } from './format/insertEntity';
