+++ packages/roosterjs-editor-types/lib/enum/ClearFormatMode.ts
@@ -1,0 +1,19 @@
+/**
+ * Represents the strategy to clear the format of the current editor selection
+ */
+export const enum ClearFormatMode {
+    /**
+     * Inline format. Remove text format.
+     */
+    Inline,
+
+    /**
+     * BLock format. Remove text and structure format of the block.
+     */
+    Block,
+
+    /**
+     * Detect Inline or Block format based on the current editor selectior.
+     */
+    AutoDetect,
+}
